moved the color sheet out to a component
